tech(api): move scoring routes in evaluation bounded context
