+++ api/src/certification/evaluation/application/scoring-configuration-route.js
