+++ api/src/certification/evaluation/infrastructure/serializers/scoring-and-capacity-simulator-report-serializer.js
