+++ api/src/certification/evaluation/routes.js
@@ -3,6 +3,8 @@ import * as certificationAdminRoutes from './application/certification-admin-rou
 import * as certificationCourseRoutes from './application/certification-course-route.js';
 import * as certificationRescoringRoutes from './application/certification-rescoring-route.js';
 import * as companionAlert from './application/companion-alert-route.js';
+import * as scoringAndCapacitySimulator from './application/scoring-and-capacity-simulator-route.js';
+import * as scoringConfiguration from './application/scoring-configuration-route.js';
 
 const certificationEvaluationRoutes = [
   companionAlert,
@@ -10,6 +12,8 @@ const certificationEvaluationRoutes = [
   certificationAdminRoutes,
   certificationRescoringRoutes,
   scenarioSimulatorRoutes,
+  scoringAndCapacitySimulator,
+  scoringConfiguration,
 ];
 
 export { certificationEvaluationRoutes };
