--- api/src/certification/scoring/routes.js
@@ -1,6 +1,0 @@
-import * as scoringAndCapacitySimulator from './application/scoring-and-capacity-simulator-route.js';
-import * as scoringConfiguration from './application/scoring-configuration-route.js';
-
-const scoringRoutes = [scoringConfiguration, scoringAndCapacitySimulator];
-
-export { scoringRoutes };
